+++ mangaScraper/Docs/ChangeLog.md
@@ -1,5 +1,5 @@
 # V 0.5
- - Added mangasearch.search()
+ - Added mangasearch.search() ([Documentation here](API.md##Search))
  - Readme changes
  - Changed parallelizeUpdateRequests to use promise.allSettled instead of promise.all
  - Fixed issue with error handling in distributeSearchRequest and distributeUpdateRequest